COSMIT `bark.h` as C header (#102)
