+++ encodec.h
@@ -129,7 +129,7 @@ struct encodec_model {
 };
 
 
-bool encodec_model_load(const std::string& fname, encodec_model& model);
+int encodec_model_load(const std::string& fname, encodec_model& model);
 
 struct ggml_tensor * encodec_quantizer_decode_eval(
                         struct ggml_context * ctx0,
