+++ examples/main.cpp
@@ -1,6 +1,24 @@
 #include "ggml.h"
 #include "bark.h"
 
+#include <tuple>
+
+std::tuple<struct bark_model *, struct bark_context *> bark_init_from_params(bark_params & params) {
+    bark_model * model = bark_load_model_from_file(params.model.c_str());
+    if (model == NULL) {
+        fprintf(stderr, "%s: error: failed to load model '%s'\n", __func__, params.model.c_str());
+        return std::make_tuple(nullptr, nullptr);
+    }
+
+    bark_context * bctx = bark_new_context_with_model(model);
+    if (bctx == NULL) {
+        fprintf(stderr, "%s: error: failed to create context with model '%s'\n", __func__, params.model.c_str());
+        bark_free_model(model);
+        return std::make_tuple(nullptr, nullptr);
+    }
+
+    return std::make_tuple(model, bctx);
+}
 
 int main(int argc, char **argv) {
     ggml_time_init();
@@ -8,14 +26,17 @@ int main(int argc, char **argv) {
 
     bark_params params;
 
-    if (bark_params_parse(argc, argv, params) == false) {
+    if (bark_params_parse(argc, argv, params) > 0) {
+        fprintf(stderr, "%s: Could not parse arguments\n", __func__);
         return 1;
     }
 
     int64_t t_load_us = 0;
     int64_t t_eval_us = 0;
 
-    bark_model model;
+    bark_context * bctx;
+    bark_model * model;
+
     std::string fname = "./ggml_weights";
 
     if (!params.model.empty()) {
@@ -23,23 +44,9 @@ int main(int argc, char **argv) {
     }
 
     // load the model
-    {
-        const int64_t t_start_us = ggml_time_us();
-
-        if(!bark_model_load(fname, model)) {
-            fprintf(stderr, "%s: failed to load model from '%s'\n", __func__, fname.c_str());
-            return 1;
-        }
-
-        t_load_us = ggml_time_us() - t_start_us;
-    }
-
-    // create a context
-    bark_context * ctx = bark_new_context_with_model(&model);
-    if (ctx == nullptr) {
-        fprintf(stderr, "%s: failed to create context\n", __func__);
-        return 1;
-    }
+    const int64_t t_start_us = ggml_time_us();
+    std::tie(model, bctx) = bark_init_from_params(params);
+    t_load_us = ggml_time_us() - t_start_us;
 
     printf("\n");
 
@@ -49,7 +56,7 @@ int main(int argc, char **argv) {
     }
 
     const int64_t t_eval_us_start = ggml_time_us();
-    bark_generate_audio(ctx, prompt.data(), params.dest_wav_path, params.n_threads);
+    bark_generate_audio(bctx, prompt.data(), params.dest_wav_path, params.n_threads);
     t_eval_us = ggml_time_us() - t_eval_us_start;
 
     // report timing
@@ -62,7 +69,7 @@ int main(int argc, char **argv) {
         printf("%s:    total time = %8.2f ms\n", __func__, (t_main_end_us - t_main_start_us)/1000.0f);
     }
 
-    bark_free(ctx);
+    bark_free(bctx);
 
     return 0;
 }
