+++ tests/test-forward-fine.cpp
@@ -22,7 +22,7 @@ int main() {
 
     bark_model model;
 
-    if(!gpt_model_load(fname, model.fine_model)) {
+    if (gpt_model_load(fname, model.fine_model) > 0) {
         fprintf(stderr, "%s: invalid model file '%s'\n", __func__, fname.c_str());
         return 1;
     }
